requirements.txt
# --- Web Framework & API ---
Django==5.0.4
djangorestframework==3.15.1

# --- Security & Environment ---
python-dotenv==1.0.1

# --- Database & Caching ---
psycopg2-binary==2.9.9
django-redis==5.4.0

# --- Core ML Engine (DAL & Anti-Distillation) ---
numpy==1.26.4
scikit-learn==1.4.2
torch==2.3.0
